Move start* methods out of Terminal and into their respective commands
Terminal now contains primarily just central Terminal stuff, and not
implementation for individual parts.
