+++ src/Terminal/commands/backdoor.ts
@@ -2,8 +2,14 @@ import { Terminal } from "../../Terminal";
 import type { TerminalAction } from "../TerminalAction";
 import { Player } from "@player";
 import { BaseServer } from "../../Server/BaseServer";
+import { HacknetServer } from "../../Hacknet/HacknetServer";
 import { Server } from "../../Server/Server";
 import { DarknetServer } from "../../Server/DarknetServer";
+import { Page } from "../../ui/Router";
+import { Router } from "../../ui/GameRoot";
+import { Engine } from "../../engine";
+import { SpecialServers } from "../../Server/data/SpecialServers";
+import { calculateHackingTime } from "../../Hacking";
 
 export function backdoor(args: (string | number | boolean)[], server: BaseServer): undefined | TerminalAction {
   if (args.length !== 0) {
@@ -38,5 +44,27 @@ export function backdoor(args: (string | number | boolean)[], server: BaseServer
     );
   }
 
-  return Terminal.startBackdoor();
+  // Backdoor should take the same amount of time as hack
+  if (server instanceof HacknetServer) {
+    Terminal.error("Cannot backdoor this kind of server");
+    return;
+  }
+  if (!(server instanceof Server || server instanceof DarknetServer)) {
+    throw new Error("server should be normal server");
+  }
+  return Terminal.timedAction(calculateHackingTime(server, Player) / 4, "backdoor", () => {
+    server.backdoorInstalled = true;
+    if (SpecialServers.WorldDaemon === server.hostname) {
+      if (Player.bitNodeN == null) {
+        Player.bitNodeN = 1;
+      }
+      Router.toPage(Page.BitVerse, { flume: false, quick: false });
+      return;
+    }
+    // Manunally check for faction invites
+    Engine.Counters.checkFactionInvitations = 0;
+    Engine.checkCounters();
+
+    Terminal.print(`Backdoor on '${server.hostname}' successful!`);
+  });
 }
