+++ src/Terminal/commands/run.ts
@@ -1,6 +1,9 @@
 import { Terminal } from "../../Terminal";
+import { Player } from "@player";
 import type { TerminalAction } from "../TerminalAction";
 import { BaseServer } from "../../Server/BaseServer";
+import { GetServer } from "../../Server/AllServers";
+import { CodingContractResult } from "../../CodingContract/Contract";
 import { runScript } from "./runScript";
 import { runProgram } from "./runProgram";
 import { hasScriptExtension } from "../../Paths/ScriptFilePath";
@@ -24,7 +27,71 @@ export function run(args: (string | number | boolean)[], server: BaseServer): un
     runScript(path, args, server);
     return;
   } else if (hasContractExtension(path)) {
-    Terminal.runContract(path).catch((error) => {
+    (async () => {
+      // There's already an opened contract
+      if (Terminal.contractOpen) {
+        return Terminal.error("There's already a Coding Contract in Progress");
+      }
+
+      const server = Player.getCurrentServer();
+      const contract = server.getContract(path);
+      if (!contract) {
+        return Terminal.error("No such contract");
+      }
+
+      Terminal.contractOpen = true;
+      const promptResult = await contract.prompt();
+
+      // Get a new copy of the server, in case it changed while the prompt was open
+      const postPromptServer = GetServer(server.hostname);
+
+      // Check if the contract still exists by the time the promise is fulfilled
+      if (postPromptServer?.getContract(path) == null) {
+        Terminal.contractOpen = false;
+        return Terminal.error("Contract no longer exists (Was it solved by a script?)");
+      }
+
+      switch (promptResult.result) {
+        case CodingContractResult.Success:
+          if (contract.reward !== null) {
+            const reward = Player.gainCodingContractReward(
+              contract.reward,
+              contract.getDifficulty(),
+              contract.rewardScaling,
+            );
+            Terminal.print(`Contract SUCCESS - ${reward}`);
+          }
+          server.removeContract(contract);
+          break;
+        case CodingContractResult.InvalidFormat:
+          Terminal.error(
+            `Contract FAILED - ${
+              promptResult.message ?? `The answer is not in the right format for contract '${contract.type}'`
+            }`,
+          );
+          break;
+        case CodingContractResult.Failure:
+          ++contract.tries;
+          if (contract.tries >= contract.getMaxNumTries()) {
+            Terminal.error("Contract FAILED - Contract is now self-destructing");
+            const solution = contract.getAnswer();
+            if (solution !== null) {
+              Terminal.error(`Coding Contract solution was: ${solution}`);
+            }
+            server.removeContract(contract);
+          } else {
+            Terminal.error(`Contract FAILED - ${contract.getMaxNumTries() - contract.tries} tries remaining`);
+          }
+          break;
+        case CodingContractResult.Cancelled:
+          Terminal.print("Contract cancelled");
+          break;
+        default: {
+          const __: never = promptResult.result;
+        }
+      }
+      Terminal.contractOpen = false;
+    })().catch((error) => {
       console.error(error);
       Terminal.error(`Cannot run contract ${path} on ${server.hostname}. Error: ${error}.`);
     });
